+++ jest.config.js
@@ -1,7 +1,5 @@
 module.exports = {
     clearMocks : true,
-    globalSetup:'./test/integration-tests/test-server/setup.ts',
-    globalTeardown:'./test/integration-tests/test-server/teardown.ts',
     moduleFileExtensions: [
         "ts",
         "js",