+++ test/integration-tests/test-server/run-test.ts
@@ -1,19 +1,18 @@
-import { initConnection } from '@enigmatis/mongo-driver';
+import { closeConnection, initConnection } from '@enigmatis/mongo-driver';
 import { config } from 'dotenv';
 import * as mongoose from 'mongoose';
-import { prepareDb } from './db-preparations';
 import { logger, server } from './server';
-
-async function init() {
+export async function init() {
     config();
-    logger.debug('JEST global setup');
+    jest.setTimeout(15000);
     const connectionString =
         'mongodb+srv://appusr:apppassword@cluster0-jotpj.mongodb.net/test?retryWrites=true';
-
     await initConnection({ connectionString, waitUntilReconnectInMs: 5000 }, logger as any);
     await mongoose.connection.db.dropDatabase();
-    await prepareDb();
     server.start();
 }
 
-module.exports = init;
+export async function finish() {
+    await closeConnection();
+    await server.stop();
+}